Add a station-stratified, continuous firefighter constraint
`constraint_factor_ff` drew globally from df_skills, so the cut fell unevenly
across stations by luck alone: at factor 2 per-station retention ranged from
34% to 100% on a nominal 50%, a 66-point spread that shifts with the seed and
hits the small stations hardest. Part of any measured effect was the draw
rather than the constraint. It was also an integer, offering only 100/50/33%,
and ff=1 -> 2 multiplies rupture_ff by ~9.5 with nothing in between.

Passing a fraction now samples within each station instead, holding every one
at the same share and narrowing the spread to ~20 points; the residue is
rounding on administrative units of one or two people. An int keeps the
historical global draw untouched, so runs made before this stay reproducible.

Stations come from `planning`, the only record of where anyone serves --
df_skills has no station column. 88% of firefighters appear in exactly one
station and the median dominant share is 100%, so attaching each to the
station holding most of their slots is faithful rather than an arbitrary
tie-break. The 1,531 never-rostered form their own stratum, sampled at the
same ratio, so exempting them cannot quietly soften the constraint.

Not yet used for a training run: at ff=2 the regime is already saturated
(rupture_ff 94,399 for 86,032 vehicles required), though strategies do
separate there -- best beats random by 13.4%, against ~0% at ff=1.

+++ agent_run_explainable.py
@@ -73,7 +73,7 @@ if __name__ == "__main__":
     parser.add_argument("--reward_weights", type=str, help="JSON file with reward weights")
     parser.add_argument("--save_metrics_as", type=str, default="dic_indic_agent", help="save metrics as")
     parser.add_argument("--constraint_factor_veh", type=int, default=1, help="size of available vehicles in Z1. factor 1 is 100%%, factor 3 is 33%%")
-    parser.add_argument("--constraint_factor_ff", type=int, default=1, help="size of available firefighters. factor 1 is 100%%, factor 3 is 33%%")
+    parser.add_argument("--constraint_factor_ff", type=float, default=1, help="size of available firefighters. factor 1 is 100%%, factor 3 is 33%%")
     parser.add_argument("--seed", type=int, default=DEFAULT_SEED, help="seed for the environment downsampling draws")
     parser.add_argument("--resume", action='store_true', help="Resume from the checkpoint written by --checkpoint_name")
     parser.add_argument("--checkpoint_name", type=str, default=None, help="Resumable checkpoint file (default: <model_name>.ckpt)")
